requirements.txt
# Django backend + API
django>=5.2.0
djangorestframework>=3.16.0
django-cors-headers>=4.4.0
gunicorn>=22.0.0

# Async jobs + realtime
celery>=5.4.0
redis>=5.0.0
channels>=4.1.0
channels-redis>=4.2.0
daphne>=4.1.0

# Database drivers
psycopg[binary]>=3.2.0

# Config
python-dotenv>=1.0.0
